+++ src/commands/sysinfo.js
@@ -7,6 +7,7 @@ const { GLOBAL_NAME } = require("../variables/constants");
 
 // Get functions
 const _convertSize = require("../functions/convSize");
+const _friendlyOS = require("../functions/friendlyOS");
 const _fatalError = require("../functions/fatalError");
 
 /**
@@ -65,25 +66,6 @@ const _determineColor = (mem) => {
     return chalk.red(mem);
   } else {
     return chalk.green(mem);
-  }
-};
-
-/**
- * Get the friendly version of the OS from `os.type()`.
- *
- * @returns The friendly name.
- */
-const _friendlyOS = () => {
-  switch (os.type()) {
-    // Darwin is another name for 'macOS', which many don't know
-    case "Darwin":
-      return "macOS";
-    // BubbleOS does not work on Windows 9x :)
-    case "Windows_NT":
-      return "Windows";
-    // If it does not match, return the default (in the case of Linux, for example)
-    default:
-      return os.type();
   }
 };